refactor(tts): extract summarization logic into tts-summarize.ts
Move summarizeText, resolveSummaryModelRef, and related types from
tts.ts into a dedicated tts-summarize.ts module. This brings tts.ts
from 1026 lines down to 902, well under the ~1000 line target.

Provider implementations (elevenLabsTTS, openaiTTS, edgeTTS) were
already extracted into tts-providers.ts by a prior commit.

All 34 tests continue to pass.
